requirements.txt
fastapi
uvicorn
requests
pandas
pypdf
python-docx
pywin32
openpyxl
python-dotenv